Put screenInit() in base class
