+++ source/application.h
@@ -34,15 +34,23 @@ class Application {
         void SetPath(std::string_view path);
 
     protected:
-        virtual appscreen screenInit() = 0;
+        appscreen screenInit() ;
         virtual appscreen screenIpSelection() = 0;
         virtual appscreen screenSendInput() = 0;
+        virtual void printHeader() = 0;
         virtual void scanPads() = 0;
+        virtual bool isHOMEHeld() = 0;
+        virtual bool isHOMEUp() = 0;
+        virtual bool isHOMEDown() = 0;
 
         static bool exitApp;
 
         GRRLIB_texImg *img_font{nullptr};
         appscreen screenId{appscreen::initapp};
+
+        // Screen IP Selection
+        std::array<std::uint8_t, 4> IP{192, 168, 1, 100};
+        std::uint16_t port{4242};
         std::string pathini{};
 
     private:
